+++ Runtime/MeshGeneration/MeshGeneratorComponent.cs
@@ -11,7 +11,7 @@ namespace MeshGeneration
         [field: SerializeField] public  MeshRenderer      MeshRendererComponent        { get; private set; }
         [field: SerializeField] public  Material          MaterialToUseForMeshRenderer { get; private set; }
         [field: SerializeField] public  string            MeshName                     { get; private set; } = "GeneratedMesh";
-        [field: SerializeField] public  MeshGeneratorBase Generator                    { get; private set; }
+        [field: SerializeField] public  MeshGenerator Generator                    { get; private set; }
 
         private Mesh m_mesh;
         public bool m_regenMesh;