+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
 
-## [1.0.0] - 2021-06-24
+## [1.1.0] - 2022-06-28
+Moved mesh related members from SharedSquareGridGenerator.cs into MeshGenerationBase.cs
+Renamed method CompleteJobs() to FinishMeshCalculations()
+Renamed SharedSquareGridGenerator to SquareGridGenerator
+Renamed MeshGeneratorBase to MeshGenerator
+
+## [1.0.0] - 2022-06-24
 This is the first release of the Mesh Generation API package.